Moved audio-player to its own folder
