+++ src/SIL.XForge.Scripture/ClientApp/src/app/shared/audio/audio-player.ts
